requirements.txt
alabaster==0.7.9
awscli==1.11.21
Babel==2.3.4
botocore==1.4.78
click==6.6
colorama==0.3.7
coverage==4.2
docutils==0.12
flake8==3.2.1
imagesize==0.7.1
Jinja2==2.8
jmespath==0.9.0
lockfile==0.12.2
luigi==2.3.3
MarkupSafe==0.23
mccabe==0.5.2
numpy==1.11.2
pandas==0.19.1
pyasn1==0.1.9
pycodestyle==2.2.0
pyflakes==1.3.0
Pygments==2.1.3
python-daemon==2.1.2
python-dateutil==2.6.0
python-dotenv==0.6.1
pytz==2016.7
PyYAML==3.12
rsa==3.4.2
s3transfer==0.1.9
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.9
tornado==4.4.2
